Lots of fixes for floating point handling; version now at 1.11.0.
Some of the following changes are ggn's, and some are mine:

 - When tokenizing floats we need to store them using a double pointer
 - PTR union needed a (double *)
 - Major changes to float depositing in eagen0.c
 - Reverted the changes in expr.c so at least floats are processed by
   expr() and friends
 - SYM svalue needs to be 64 bits
 - When EQUing a float symbol don't chop off the upper 32 bits from eval
 - Added fltpoint.{c,h} in order to properly create IEEE-754 floating
   point and Motorola extended numbers
 - Fixed float evaluations in evexpr()
 - Fixed floating point depositions in direct.c (in d_dc())
 - Upped the BSD image limit in object.c to 8MB for crazy people making
   6MB Jaguar ROMs (will need a real fix at some point)

+++ procln.c
@@ -575,7 +575,7 @@ When checking to see if it's already been equated, issue a warning.
 		}
 
 		sy->sattr |= eattr | EQUATED;	// Symbol inherits value and attributes
-		sy->svalue = (uint32_t)eval;
+		sy->svalue = eval;
 
 		if (list_flag)					// Put value in listing
 			listvalue((uint32_t)eval);
